+++ previous-versions/files/src/codatfiles/models/operations/list_files.py
@@ -3,10 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import file as shared_file
-from ..shared import schema as shared_schema
-from codatfiles import utils
-from dataclasses_json import Undefined, dataclass_json
+from ...models.shared import file as shared_file
 from typing import List, Optional
 
 
@@ -18,39 +15,15 @@ class ListFilesRequest:
 
 
 
-@dataclass_json(undefined=Undefined.EXCLUDE)
-@dataclasses.dataclass
-class ListFilesErrorMessage:
-    r"""You are using an outdated API key or a key not associated with that resource."""
-    can_be_retried: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('canBeRetried'), 'exclude': lambda f: f is None }})
-    r"""`True` if the error occurred transiently and can be retried."""
-    correlation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('correlationId'), 'exclude': lambda f: f is None }})
-    r"""Unique identifier used to propagate to all downstream services and determine the source of the error."""
-    detailed_error_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('detailedErrorCode'), 'exclude': lambda f: f is None }})
-    r"""Machine readable error code used to automate processes based on the code returned."""
-    error: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('error'), 'exclude': lambda f: f is None }})
-    r"""A brief description of the error."""
-    service: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('service'), 'exclude': lambda f: f is None }})
-    r"""Codat's service the returned the error."""
-    status_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('statusCode'), 'exclude': lambda f: f is None }})
-    r"""The HTTP status code returned by the error."""
-    
-
-
-
 @dataclasses.dataclass
 class ListFilesResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[ListFilesErrorMessage] = dataclasses.field(default=None)
-    r"""You are using an outdated API key or a key not associated with that resource."""
     files: Optional[List[shared_file.File]] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
-    schema: Optional[shared_schema.Schema] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""